requirements.txt
pdfplumber
python-docx
fastapi
uvicorn
python-multipart
pdfplumber
python-dotenv
google-generativeai

